+++ src/dipdup/indexes/tezos_tzkt_operations/matcher.py
@@ -34,7 +34,7 @@ class OperationSubgroup:
     entrypoints: set[str | None]
 
 
-OperationsHandlerArgumentU = TzktTransaction | TzktOrigination | TzktOperationData | None
+OperationsHandlerArgumentU = TzktTransaction[Any, Any] | TzktOrigination[Any] | TzktOperationData | None
 MatchedOperationsT = tuple[OperationSubgroup, TzktOperationsHandlerConfigU, deque[OperationsHandlerArgumentU]]