+++ CHANGELOG.md
@@ -8,7 +8,13 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ### Added
 
+- evm.node: Requesting logs for multiple blocks at once for EVM nodes to improve performance
+- evm.subsquid: Added metrics for Subsquid Cloud deploy.
 - tezos.tzkt: Added support for Etherlink smart rollups (`sr1…` addresses).
+
+### Security
+
+- deps: Updated PyArrow to 14.0.1 to fix [CVE-2023-47248](https://github.com/advisories/GHSA-5wvp-7f3h-6wmm)
 
 ## [7.1.1] - 2023-11-07